--- a/ecmascript/parser/src/lexer/mod.rs
+++ b/ecmascript/parser/src/lexer/mod.rs
@@ -769,17 +769,6 @@ impl<'a, I: Input> Lexer<'a, I> {
         let content_start = self.cur_pos();
 
         while let Some(c) = self.cur() {
-            if !escaped {
-                let chunk = self.input.uncons_while(|c| {
-                    !c.is_line_break() && c != '[' && c != ']' && c != '/' && c != '\\'
-                });
-
-                if !chunk.is_empty() {
-                    content.push_str(chunk);
-                    continue;
-                }
-            }
-
             // This is ported from babel.
             // Seems like regexp literal cannot contain linebreak.
             if c.is_line_break() {