--- a/ecmascript/parser/src/lexer/number.rs
+++ b/ecmascript/parser/src/lexer/number.rs
@@ -2,10 +2,9 @@
 //!
 //!
 //! See https://tc39.github.io/ecma262/#sec-literals-numeric-literals
-
 use super::*;
 use crate::error::SyntaxError;
-use std::fmt::Display;
+use std::{fmt::Write, iter::FusedIterator};
 
 impl<'a, I: Input> Lexer<'a, I> {
     /// Reads an integer, octal integer, or floating-point number
@@ -48,21 +47,24 @@ impl<'a, I: Input> Lexer<'a, I> {
                     // strict mode hates non-zero decimals starting with zero.
                     // e.g. 08.1 is strict mode violation but 0.1 is valid float.
 
-                    let s = format!("{}", val); // TODO: Remove allocation.
+                    if val.fract() < 1e-10 {
+                        let d = digits(val.round() as u64, 10);
 
-                    // if it contains '8' or '9', it's decimal.
-                    if s.contains('8') || s.contains('9') {
-                        if self.ctx.strict {
-                            self.error(start, SyntaxError::LegacyDecimal)?
+                        // if it contains '8' or '9', it's decimal.
+                        if d.clone().any(|v| v == 8 || v == 9) {
+                            if self.ctx.strict {
+                                self.error(start, SyntaxError::LegacyDecimal)?
+                            }
+                        } else {
+                            // It's Legacy octal, and we should reinterpret value.
+                            let val = u64::from_str_radix(&val.to_string(), 8)
+                                .expect("Does this can really happen?");
+                            let val = val
+                                .to_string()
+                                .parse()
+                                .expect("failed to parse numeric value as f64");
+                            return self.make_legacy_octal(start, val);
                         }
-                    } else {
-                        // It's Legacy octal, and we should reinterpret value.
-                        let val = u64::from_str_radix(&format!("{}", val), 8)
-                            .expect("Does this can really happen?");
-                        let val = format!("{}", val)
-                            .parse()
-                            .expect("failed to parse numeric value as f64");
-                        return self.make_legacy_octal(start, val);
                     }
                 }
             }
@@ -86,17 +88,14 @@ impl<'a, I: Input> Lexer<'a, I> {
 
             // Read numbers after dot
             let dec_val = self.read_int(10, 0, &mut Raw(None))?;
+            let mut s = String::new();
+            write!(s, "{}.", val).unwrap();
 
-            let dec: &dyn Display = match dec_val {
-                Some(ref n) => n,
-                // "0.", "0.e1" is valid
-                None => &"",
-            };
+            if let Some(ref n) = dec_val {
+                write!(s, "{}", n).unwrap();
+            }
 
-            // TODO
-            val = format!("{}.{}", val, dec)
-                .parse()
-                .expect("failed to parse float using rust's impl");
+            val = s.parse().expect("failed to parse float using rust's impl");
         }
 
         // Handle 'e' and 'E'
@@ -305,6 +304,46 @@ impl<'a, I: Input> Lexer<'a, I> {
 
         return Ok(val);
     }
+}
+
+fn digits(value: u64, radix: u64) -> impl Iterator<Item = u64> + Clone + 'static {
+    debug_assert!(radix > 0);
+
+    #[derive(Clone, Copy)]
+    struct Digits {
+        n: u64,
+        divisor: u64,
+    }
+
+    impl Digits {
+        fn new(n: u64, radix: u64) -> Self {
+            let mut divisor = 1;
+            while n >= divisor * radix {
+                divisor *= radix;
+            }
+
+            Digits { n, divisor }
+        }
+    }
+
+    impl Iterator for Digits {
+        type Item = u64;
+
+        fn next(&mut self) -> Option<u64> {
+            if self.divisor == 0 {
+                None
+            } else {
+                let v = Some(self.n / self.divisor);
+                self.n %= self.divisor;
+                self.divisor /= 10;
+                v
+            }
+        }
+    }
+
+    impl FusedIterator for Digits {}
+
+    Digits::new(value, radix)
 }
 
 #[cfg(test)]