Fixes #3.
Fixes #458.

```
test angular          ... bench:  31,612,004 ns/iter (+/- 601,832) = 22 MB/s
test angular_ts       ... bench:  33,168,604 ns/iter (+/- 483,394) = 21 MB/s
test backbone         ... bench:   4,463,825 ns/iter (+/- 216,895) = 13 MB/s
test backbone_ts      ... bench:   4,718,108 ns/iter (+/- 96,289) = 12 MB/s
test colors           ... bench:      56,156 ns/iter (+/- 6,190) = 20 MB/s
test colors_ts        ... bench:      58,322 ns/iter (+/- 7,543) = 19 MB/s
test jquery           ... bench:  25,452,922 ns/iter (+/- 562,472) = 10 MB/s
test jquery_mobile    ... bench:  40,899,168 ns/iter (+/- 838,919) = 11 MB/s
test jquery_mobile_ts ... bench:  44,040,351 ns/iter (+/- 14,766,527) = 10 MB/s
test jquery_ts        ... bench:  26,953,415 ns/iter (+/- 809,036) = 9 MB/s
test mootools         ... bench:  20,687,642 ns/iter (+/- 730,233) = 7 MB/s
test mootools_ts      ... bench:  21,741,575 ns/iter (+/- 534,243) = 7 MB/s
test underscore       ... bench:   3,848,639 ns/iter (+/- 268,541) = 11 MB/s
test underscore_ts    ... bench:   4,124,005 ns/iter (+/- 246,357) = 10 MB/s
test yui              ... bench:  18,678,015 ns/iter (+/- 616,555) = 18 MB/s
test yui_ts           ... bench:  19,841,452 ns/iter (+/- 862,464) = 17 MB/s
```